gui update, piece value scaled basic ai using ab-pruning minimax with adjustable depth, board evaluator only considers piece value for now

+++ src/main/java/com/chess/engine/pieces/Piece.java
@@ -60,7 +60,7 @@ public abstract class Piece {
         return result;
     }
 
-    public abstract double getValue();
+    public abstract int getValue();
 
     public abstract Collection<Move> calculateLegalMoves(Board board);
